Add logical package structure for websockets
