--- celery_progress/routing.py
@@ -1,7 +1,0 @@
-from django.conf.urls import url
-
-from . import consumers
-
-websocket_urlpatterns = [
-    url(r'^ws/progress/(?P<task_id>[\w-]+)/$', consumers.ProgressConsumer),
-]
