+++ celery_progress/static/celery_progress/websockets.js
